fix: move config map check to trainablemodule
